+++ identity/src/main/java/bisq/identity/IdentityService.java
@@ -32,9 +32,9 @@ import bisq.network.p2p.node.Node;
 import bisq.persistence.Persistence;
 import bisq.persistence.PersistenceClient;
 import bisq.persistence.PersistenceService;
-import bisq.security.PubKey;
 import bisq.security.keys.KeyBundle;
 import bisq.security.keys.KeyBundleService;
+import bisq.security.keys.PubKey;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Streams;
 import lombok.Getter;
