+++ identity/src/test/java/bisq/identity/IdentityServiceTest.java
@@ -24,8 +24,8 @@ import bisq.network.common.TransportType;
 import bisq.network.identity.NetworkId;
 import bisq.network.p2p.node.Node;
 import bisq.persistence.PersistenceService;
-import bisq.security.PubKey;
 import bisq.security.keys.KeyBundleService;
+import bisq.security.keys.PubKey;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
