+++ network/network-identity/src/main/java/bisq/network/identity/NetworkIdWithKeyPair.java
@@ -17,7 +17,7 @@
 
 package bisq.network.identity;
 
-import bisq.security.PubKey;
+import bisq.security.keys.PubKey;
 import lombok.Getter;
 import lombok.ToString;
