+++ network/network/src/main/java/bisq/network/p2p/ServiceNode.java
@@ -43,8 +43,8 @@ import bisq.network.p2p.services.data.inventory.InventoryService;
 import bisq.network.p2p.services.peergroup.BanList;
 import bisq.network.p2p.services.peergroup.PeerGroupManager;
 import bisq.persistence.PersistenceService;
-import bisq.security.PubKey;
 import bisq.security.keys.KeyBundleService;
+import bisq.security.keys.PubKey;
 import com.runjva.sourceforge.jsocks.protocol.Socks5Proxy;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
