+++ network/network/src/main/java/bisq/network/p2p/services/data/storage/auth/RefreshAuthenticatedDataRequest.java
@@ -23,8 +23,8 @@ import bisq.common.validation.NetworkDataValidation;
 import bisq.network.p2p.services.data.DataRequest;
 import bisq.network.p2p.services.data.storage.MetaData;
 import bisq.security.DigestUtil;
-import bisq.security.KeyGeneration;
 import bisq.security.SignatureUtil;
+import bisq.security.keys.KeyGeneration;
 import com.google.protobuf.ByteString;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
