+++ network/network/src/main/java/bisq/network/p2p/services/data/storage/auth/authorized/AuthorizedData.java
@@ -21,8 +21,8 @@ import bisq.common.encoding.Hex;
 import bisq.common.validation.NetworkDataValidation;
 import bisq.network.p2p.services.data.storage.DistributedData;
 import bisq.network.p2p.services.data.storage.auth.AuthenticatedData;
-import bisq.security.KeyGeneration;
 import bisq.security.SignatureUtil;
+import bisq.security.keys.KeyGeneration;
 import com.google.protobuf.ByteString;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
