+++ network/network/src/main/java/bisq/network/p2p/services/data/storage/mailbox/AddMailboxRequest.java
@@ -21,8 +21,8 @@ import bisq.common.util.MathUtils;
 import bisq.common.validation.NetworkDataValidation;
 import bisq.network.p2p.services.data.AddDataRequest;
 import bisq.security.DigestUtil;
-import bisq.security.KeyGeneration;
 import bisq.security.SignatureUtil;
+import bisq.security.keys.KeyGeneration;
 import com.google.protobuf.ByteString;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
