+++ network/network/src/main/java/bisq/network/p2p/services/data/storage/mailbox/MailboxSequentialData.java
@@ -19,7 +19,7 @@ package bisq.network.p2p.services.data.storage.mailbox;
 
 import bisq.common.proto.Proto;
 import bisq.common.validation.NetworkDataValidation;
-import bisq.security.KeyGeneration;
+import bisq.security.keys.KeyGeneration;
 import com.google.protobuf.ByteString;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
