+++ oracle_node/src/main/java/bisq/oracle_node/bisq1_bridge/Bisq1BridgeService.java
@@ -41,8 +41,8 @@ import bisq.persistence.Persistence;
 import bisq.persistence.PersistenceClient;
 import bisq.persistence.PersistenceService;
 import bisq.security.DigestUtil;
-import bisq.security.KeyGeneration;
 import bisq.security.SignatureUtil;
+import bisq.security.keys.KeyGeneration;
 import bisq.user.reputation.data.AuthorizedAccountAgeData;
 import bisq.user.reputation.data.AuthorizedBondedReputationData;
 import bisq.user.reputation.data.AuthorizedProofOfBurnData;
